+++ RELEASES.md
@@ -1,3 +1,12 @@
+# Release 0.1.3 (2021-06-23)
+
+- Bezier arcs and rounded corners now use
+   a better approach using a matched polynomial
+   to derive 'lambda'; this is now tested in
+   regression to a good degree of accuracy.
+
+**Contributors**: @atthecodeface
+
 # Release 0.1.0 (2021-06-22)
 
 - Publishing on crates.io for the first time